+++ qiskit/synthesis/two_qubits/__init__.py
@@ -11,3 +11,5 @@
 # that they have been altered from the originals.
 
 """Module containing two-qubit unitary synthesis methods."""
+
+from .two_qubit_decompose import TwoQubitBasisDecomposer, two_qubit_cnot_decompose
