+++ qiskit/synthesis/two_qubits/two_qubit_decompose.py
@@ -39,7 +39,7 @@ from qiskit.circuit import QuantumRegister, QuantumCircuit, Gate
 from qiskit.circuit.library.standard_gates import CXGate, RXGate, RYGate, RZGate
 from qiskit.exceptions import QiskitError
 from qiskit.quantum_info.operators import Operator
-from qiskit.quantum_info.synthesis.weyl import weyl_coordinates, transform_to_magic_basis
+from qiskit.synthesis.two_qubits.weyl import weyl_coordinates, transform_to_magic_basis
 from qiskit.synthesis.one_qubit.one_qubit_decompose import (
     OneQubitEulerDecomposer,
     DEFAULT_ATOL,
