+++ test/randomized/test_synthesis.py
@@ -21,7 +21,7 @@ from qiskit.circuit import QuantumCircuit, QuantumRegister
 from qiskit.circuit.library import UnitaryGate
 from qiskit.providers.basicaer import UnitarySimulatorPy
 from qiskit.quantum_info.random import random_unitary
-from qiskit.quantum_info.synthesis.two_qubit_decompose import (
+from qiskit.synthesis.two_qubits.two_qubit_decompose import (
     two_qubit_cnot_decompose,
     TwoQubitBasisDecomposer,
     Ud,
